exo 8

+++ TP2/tp3exo8.py
@@ -1,4 +1,4 @@
-x = int(input("Entrez un nombre décimal : "))
+x = float(input("Entrez un nombre décimal : "))
 if (2<= x <3) or (0< x <=1) or (-10<= x <=2):
     print(f"{x} appartient à I")
 else:
